feat(charts): add prelim support for Ingress Class
You can set the controller's ingress class, but more design is necessary
to come up with a good way for apps to set their own ingress class (or
even for the controller chart to create the ingresses with a default
setting for ingress class.)

This is a non-breaking change but technically a feature. I can extract
it into another PR or you're free to merge it as-is.